Move the proxy server and related to `cluster-proxy-server`
The proxy server that can used MSW handlers (generated from har files)
and sits between bridge and the cluster will now reside in the
`cluster-proxy-server` directory.

It will be peer with:
  - `console-plugin` (MSW support via openshift-console context provider extension)
  - `jest` (MSW support for jest unit tests)
