+++ packages/mocks/src/browser.ts
@@ -1,5 +1,5 @@
 import { setupWorker } from 'msw';
 
-import { createDefaultHandlers } from './handlers';
+import { createDefaultHandlers } from './cluster-proxy-server/handlers';
 
 export const worker = setupWorker(...createDefaultHandlers({}));
